Backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pandas>=2.2.0
openpyxl>=3.1.0
pypdf>=5.0.0
pymupdf>=1.24.0
python-docx>=1.1.0
numpy>=2.0.0
scikit-learn>=1.5.0
sentence-transformers>=3.0.0
typer>=0.12.0
python-multipart>=0.0.9
chromadb>=0.5.0
PyPDF2
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
python-dotenv>=1.0.1
pydantic-settings>=2.3.0
azure-storage-blob>=12.0.0
torch>=2.2.0
transformers>=4.40.0
tf-keras>=2.16.0
tokenizers>=0.19.0
peft>=0.10.0
accelerate>=0.29.0
tqdm
huggingface_hub>=0.23.0
